requirements.txt
streamlit
requests
pandas
google-genai